requirements_train.txt
wheel
sagemaker==2.86.0
boto3==1.21.39
numpy==1.22.3
matplotlib==3.5.1
torchsummary==1.5.1
tqdm==4.64.0
scikit-learn==1.0.2
pandas==1.4.2
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.11.0+cu102
torchvision==0.12.0+cu102
Pillow==9.1.0